robot-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.6.1
